etl/requirements.txt
affine==2.4.0
attrs==25.4.0
certifi==2025.10.5
cftime==1.6.5
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
GeoAlchemy2==0.18.0
geopandas==1.1.1
greenlet==3.2.4
netCDF4==1.7.3
numpy==2.2.6
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rasterio==1.4.3
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
